pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "ragable"
dynamic = ["version"]
authors = [ 
  { name="datapsycho", email="mr.data.psycho@gmail.com" },
]
description = "Applciation which is ragable"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["Retrival Augmented Generation", "genAI", "OpenAI"]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.1",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "chainlit>=1.1.101",
    "tiktoken>=0.7.0",
    "cohere>=5.5.0",
]
license = {file = "LICENSE"}

[project.optional-dependencies]
backend = ["fastapi"]
frontend = ["chainlit"]
dev = ["isort", "ruff"]

[tool.hatch.build.targets.wheel]
sources = ["src"]
only-include = ["src/ragable"]

[tool.hatch.version]
path = "src/ragable/__about__.py"


[tool.flake8]
ignore = ['E231', 'E241']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 120
count = true


#[tool.pytest.ini_options]
#minversion = "6.0"
#addopts = "--no-header -ra -v" # -s for pringing in tests
#testpaths = [
#    "tests",
#]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report xml --cov-report term-missing"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "integration"
]
norecursedirs = ["src/authentifield/model_template/auto_AD/"]

[tool.coverage]
    [tool.coverage.run]
    omit = [
        # omit anything in a .local directory anywhere
        '__init__.py',
        'tests/*',
        '*/tests/*',
        # omit anything in a .venv directory anywhere
        '.venv/*'
    ]

    [tool.coverage.report]
    skip_empty = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 120